[ADD] Box2DComponent to add basic physics emulation participation and functionality to game objects. [MOD] Each scene now has an individual physics world. [MOD] The tilemap level loaded from level files now has collision functionality. [DEL] Removed the singleton pattern in favor of using the service locator pattern for globally required services.

+++ Minigin/GameObject.cpp
@@ -5,7 +5,7 @@
 
 GameObject::~GameObject()
 {
-	for (int i{}; i < m_Components.size(); i++)
+	for (size_t i{}; i < m_Components.size(); i++)
 	{
 		delete m_Components[i];
 		m_Components[i] = nullptr;
